new empty lit test file
